requirements.txt
aiohttp==3.7.4.post0
apache-beam==2.32.0
async-timeout==3.0.1
attrs==19.3.0
avro-python3==1.9.2.1
backports.entry-points-selectable==1.1.0
blinker==1.4
cached-property==1.5.1
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.0.6
Click==7.0
colorama==0.4.3
crcmod==1.7
cryptography==2.8
dbus-python==1.2.16
dill==0.3.1.1
discord.py==1.7.3
distlib==0.3.2
distro==1.4.0
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
fastavro==1.4.4
filelock==3.0.12
future==0.18.2
grpcio==1.40.0
hdfs==2.6.0
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
jsonschema==3.2.0
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
macaroonbakery==1.3.1
meson==0.53.2
more-itertools==4.2.0
multidict==5.2.0
netifaces==0.10.4
numpy==1.20.3
oauth2client==4.1.3
oauthlib==3.1.0
olefile==0.46
orjson==3.6.3
pexpect==4.6.0
Pillow==7.0.0
platformdirs==2.2.0
protobuf==3.18.0
pyarrow==4.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycups==1.9.73
pydot==1.4.2
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==3.12.0
PyNaCl==1.3.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.15.5
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.26.0
requests-unixsocket==0.2.0
rsa==4.7.2
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
systemd-python==234
texttable==1.6.2
typing-extensions==3.7.4.3
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
virtualenv==20.7.2
wadllib==1.3.3
websocket-client==0.53.0
xkit==0.0.0
xmltodict==0.12.0
yarl==1.7.0
zipp==1.0.0
